core/json: Rename json_path.rs to path.rs
